refactor: make a base python plugin
